scoping_review_agent/requirements.txt
requests
pyyaml
pandas
python-docx
pypdf

